Last commit, hopefully

+++ 0x00-hello_world/6-size.c
@@ -14,11 +14,11 @@ long along;
 long long alonglong;
 float afloat;
 
-printf("Size of a char: %zu byte(s)\n", (unsigned long) sizeof(acharacter));
+printf("Size of a char: %zu byte(s)\n", sizeof(acharacter));
 printf("size of an int: %zu byte(s)\n", sizeof(ainteger));
-printf("Size of a long int: %zu byte(s)\n", (unsigned long) sizeof(along));
-printf("Size of a long long int: %zu byte(s)\n", (unsigned long) sizeof(alonglong));
-printf("Size of a float: %zu byte(s)\n", (unsigned long) sizeof(afloat));
+printf("Size of a long int: %zu byte(s)\n", sizeof(along));
+printf("Size of a long long int: %zu byte(s)\n", sizeof(alonglong));
+printf("Size of a float: %zu byte(s)\n", sizeof(afloat));
 
 return (0);
